Refactored internal classes into their own files.
