+++ framework/Pathfinder/Pathfinder.swift
@@ -16,12 +16,15 @@ A connection to the Pathfinder service, specific to one authenticated user belon
 This is the starting point for all interactions with Pathfinder. The typical use case of this library is:
 
 1. Retrieve a Cluster, either by id or the default cluster of the application. All transports, commodities and routes within specific clusters. Additional clusters can be configured via the Pathfinder web interface.
-2. Register as a transport or create a commodity within that cluster.
+2. Register the device as a transport by creating a new transport instance or requesting transportation by create a commodity.
 3. Create a delegate object for the newly created transport or commodity.
+4. Subscribe to the vehicle, commodity or any other vehicles or commodities to receive status and location updates.
 4. Respond to the CommodityDelegate or TransportDelegate protocol methods by displaying them on a MapView and executing the appropriate business logic.
 */
 public class Pathfinder {
-  let applicationIdentifier: String
+  /// The application id that was assigned by the Pathfinder web portal when you registered your application.
+  public let applicationIdentifier: String
+
   let userCredentials: String
   let conn: PathfinderConnection
 
@@ -37,22 +40,17 @@ public class Pathfinder {
     self.conn = PathfinderConnection(applicationIdentifier: applicationIdentifier)
   }
 
-  /**
-  Retrieves the top-level cluster for the application, including references to all transports and commodities within it. The resulting object can be used to watch for events including transport and commodity updates and route assignments.
-
-  - Parameter callback:  This function will be called exactly once with the populated Cluster object once it is retrieved.
-  */
-  public func defaultCluster() -> Cluster {
+  /// Retrieves the top-level cluster for the application, including references to all transports and commodities within it. The resulting object can be used to watch for events including transport and commodity updates and route assignments.
+  public func cluster() -> Cluster {
     return Cluster(conn: conn)
   }
 
   /**
-  Retrieves a previously created cluster by id number for the application, including references to all transports and commodities within it. The resuling object can be used to watch for events including transport and commodity updates and route assignments. This method will only need to be used if your application requires Pathfinder subclusters.
+  Retrieves a cluster by path, including references to all transports and commodities within it. The resulting object can be used to watch for events including transport and commodity updates and route assignments. This method will only need to be used if your application requires Pathfinder subclusters.
 
-  - Paremeter id:        The id of the cluster to retrieve.
-  - Parameter callback:  This function will be called exactly once with the populated Cluster object once it is retrieved.
+  - Paremeter path:  The path of the application subcluster.
   */
-  public func clusterById(id: Int) -> Cluster {
-    return Cluster(conn: conn, id: id)
+  public func cluster(path: String) -> Cluster {
+    return Cluster(conn: conn, path: path)
   }
 }
