+++ framework/Pathfinder/Route.swift
@@ -65,78 +65,3 @@ public class Route {
     return nil
   }
 }
-
-/// A data object containing a commodity, a location and a field indicating pickup or dropoff.
-public class RouteAction {
-
-  /// The possible actions that can be performed on commodities by transports.
-  public enum Action {
-    case Start
-    case Pickup
-    case Dropoff
-  }
-
-  /// Whether the action is a pickup or dropoff.
-  public let action: Action
-
-  /// The commodity that is picked up or dropped off.
-  public var commodity: Commodity?
-
-  /// The location where the action is to occur.
-  public let location: CLLocationCoordinate2D
-
-  init(action: Action, location: CLLocationCoordinate2D) {
-    self.action = action
-    self.location = location
-  }
-
-  init(action: Action, commodity: Commodity, location: CLLocationCoordinate2D) {
-    self.action = action
-    self.commodity = commodity
-    self.location = location
-  }
-
-  class func parse(message: NSDictionary) -> RouteAction? {
-    if let actionString = message["action"] as? String {
-      let latitude = message["latitude"] as! Double
-      let longitude = message["longitude"] as! Double
-      let location = CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
-      print("Attempting to parse RouteAction at location \(location)")
-      if actionString == "start" {
-        return RouteAction(action: Action.Start, location: location)
-      } else if actionString == "pickup" {
-        if let commodity = Commodity.parse(message["commodity"] as! NSDictionary) {
-          return RouteAction(action: Action.Pickup, commodity: commodity, location: location)
-        }
-      } else if actionString == "dropoff" {
-        if let commodity = Commodity.parse(message["commodity"] as! NSDictionary) {
-          return RouteAction(action: Action.Dropoff, commodity: commodity, location: location)
-        }
-      }
-    }
-    print("Failed to parse RouteAction: \(message)")
-    return nil
-  }
-}
-
-class RoutedResponse {
-  class func parse(message: NSDictionary) -> RoutedResponse? {
-    if let routed = message["routed"] as? NSDictionary {
-      let model = routed["model"] as! String
-      if let routeDict = routed["route"] as? NSDictionary {
-        if let route = Route.parse(routeDict) {
-          return RoutedResponse(model: model, route: route)
-        }
-      }
-    }
-    return nil
-  }
-
-  let model: String
-  let route: Route
-
-  init(model: String, route: Route) {
-    self.model = model
-    self.route = route
-  }
-}
